parser package organization
